Refactor to hide away `useBundler` as implementation detail
